✨ feat: Introduce new features
Add commit hook creation and removal, theme variable, error code check,
and updates to scripts and components. Also, remove and reorder components,
add new components, update imports and dependencies, and make adjustments
to various files.

The changes introduce new features and improve the project's functionality.

+++ packages/lobe-commit/src/commands/Hook/HookRemove.tsx
@@ -1,22 +1,33 @@
 import { Alert, Spinner } from '@inkjs/ui';
+import { useTheme } from '@lobehub/cli-ui';
+import { Text } from 'ink';
 import fs from 'node:fs';
+import process from 'node:process';
 import { memo, useEffect, useState } from 'react';
 
 import { HOOK } from '@/constants/hook';
-import getAbsoluteHooksPath from '@/utils/getAbsoluteHooksPath';
+import getAbsoluteHooksPath, { ERROR_CODE } from '@/utils/getAbsoluteHooksPath';
 
 const HookRemove = memo(() => {
   const [loading, setLoading] = useState<boolean>(true);
+  const theme = useTheme();
   try {
     useEffect(() => {
       const hookFile = getAbsoluteHooksPath(HOOK.FILENAME);
+      if (hookFile === ERROR_CODE) process.exit(1);
       fs.unlinkSync(hookFile);
       setLoading(false);
     }, []);
     if (loading) return <Spinner label=" Loading..." />;
-    return <Alert variant="success">{` Lobe-commit hook removed successfully!`}</Alert>;
+    return (
+      <Alert variant="success">
+        {` lobe-commit hook `}
+        <Text color={theme.colorError}>removed</Text>
+        {` successfully!`}
+      </Alert>
+    );
   } catch {
-    return <Alert variant="error">{` Error: Gitmoji commit hook is not created`}</Alert>;
+    return <Alert variant="error">{` lobe-commit commit hook is not found`}</Alert>;
   }
 });
